chore: move green visualizer 2D to a separate script for future extended use
